+++ src/app/datasets/onedep/onedep.component.html
@@ -360,9 +360,7 @@
                         </tr>
                         <!-- UPLOAD FILES DONE -->
 
-                        <button class="submitDep" mat-raised-button color="primary" (click)="connectToDepositionBackend()">
-                            connect
-                        </button>
+                 
                         <button class="submitDep" mat-raised-button color="primary" (click)="onDepositClick()">
                             Start Deposition
                         </button>
